extract GrantBrandingRoutingPolicy to separate file
